// src/api/generated/model/employeeBasicDto.ts
/**
 * Generated by orval — do not edit manually.
 * MySweldo HRIS API 0.0.1
 */import type { DepartmentBasicDto } from './departmentBasicDto';
import type { EmployeeBasicDtoStatus } from './employeeBasicDtoStatus';
import type { EmployeeBasicDtoType } from './employeeBasicDtoType';
import type { PositionBasicDto } from './positionBasicDto';

export interface EmployeeBasicDto {
  id: number;
  firstName: string;
  lastName: string;
  department: DepartmentBasicDto;
  position: PositionBasicDto;
  status: EmployeeBasicDtoStatus;
  type: EmployeeBasicDtoType;
}