// src/api/generated/model/employee-basic-dto.ts
/**
 * Generated by orval — do not edit manually.
 * MySweldo HRIS API 0.0.1
 */import type { DepartmentBasicDto } from './department-basic-dto';
import type { EmployeeBasicDtoStatus } from './employee-basic-dto-status';
import type { EmployeeBasicDtoType } from './employee-basic-dto-type';
import type { PositionBasicDto } from './position-basic-dto';

export interface EmployeeBasicDto {
  id: number;
  firstName: string;
  lastName: string;
  department: DepartmentBasicDto;
  position: PositionBasicDto;
  status: EmployeeBasicDtoStatus;
  type: EmployeeBasicDtoType;
}